pyproject.toml
[tool.poetry]
name = "hardware_scraping"
version = "0.1.0"
description = ""
authors = ["Dmitry Bunin <bunin.da@phystech.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
Scrapy = "^2.5.1"
requests = "^2.27.1"
beautifulsoup4 = "^4.11.1"
python-dateutil = "^2.8.2"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
